# Reworked timers, performance tests in python