## PerformanceNodes and Timers
There are now `PerformanceNode`s that can create RAII Timers. Every node can create subnodes, which again can create Timers. Thus a tree-like structure emerges, that represents the program-flow. 

Nodes of this tree carry 
- a label
- the cumulative time
- the number of times this node has been profiled.

The root of this tree lives in the simulation object. After the `run()` it can be accessed in python.
One can acess every node of this tree in a filesystem-like manner. E.g. the following code
```python
root = simulation.performance_root()
print(root)
print(root[""])
print(root["integrator"])
print(root["integrator"].time())
print(root["integrator"].count())
print(root.keys())
```
would produce the following output
```
simulation: time 0.012835 s, count 1
	neighborList: time 0.00646 s, count 21
	forces: time 0.002056 s, count 11
	integrator: time 0.001159 s, count 10
	reactionScheduler: time 1.2e-05 s, count 10
	clearNeighborList: time 1e-06 s, count 1

simulation: time 0.012835 s, count 1
	neighborList: time 0.00646 s, count 21
	forces: time 0.002056 s, count 11
	integrator: time 0.001159 s, count 10
	reactionScheduler: time 1.2e-05 s, count 10
	clearNeighborList: time 1e-06 s, count 1

integrator: time 0.001159 s, count 10

0.0011589999999999999
10
['neighborList', 'forces', 'integrator', 'reactionScheduler', 'clearNeighborList']
```

## Performance tests
The performance tests have moved from C++ to Python. The structure is very similar. There are `Scenario`s (such as "only softcore particles uniformly distributed") that wrap one `Simulation` instance. There are `Sampler`s that sample a certain scenario along a given set of variables (e.g. number of particles).